requirements.txt
aioquic==0.9.21
aiosqlite==0.22.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.10.0
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
brotli==1.2.0
browser-cookie3==0.20.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
defusedxml==0.7.1
dnspython==2.8.0
ecdsa==0.19.1
fastapi==0.109.2
fonttools==4.61.1
fpdf2==2.8.5
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.27.0
httpx-ntlm==1.4.0
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
jeepney==0.9.0
Jinja2==3.1.6
kaitaistruct==0.11
ldap3==2.9.1
lz4==4.4.5
Mako==1.3.10
MarkupSafe==3.0.3
msgpack==1.1.2
packaging==25.0
passlib==1.7.4
pillow==12.1.0
playwright==1.55.0
publicsuffix2==2.20191221
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.6.1
pydantic_core==2.16.2
pyee==13.0.0
pylsqpack==0.3.23
pyOpenSSL==25.3.0
pyparsing==3.2.5
pyperclip==1.11.0
PySocks==1.7.1
pyspnego==0.12.0
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.31.0
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
scapy==2.7.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.27
sqlmap==1.9.12
starlette==0.36.3
tld==0.13.1
tornado==6.5.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
urwid==3.0.3
uvicorn==0.27.1
wapiti-swagger==0.1.9
wapiti3==3.0.3
wcwidth==0.2.14
Werkzeug==3.1.4
wsproto==1.2.0
yaswfp==0.9.3
zstandard==0.25.0
mysql-connector-python==8.3.0
